fixed constants

+++ src/main/java/frc/robot/Constants.java
@@ -22,7 +22,7 @@ public final class Constants {
     public static final class Manipulator {
         public static final int leaderId = 14;
         public static final int followerId = 15;
-        //TODO check the real channel
+        // TODO check the real channel
         public static final int leftChannel = 8;
         public static final int rightChannel = 9;
         public static final double kP = 0.1; //
@@ -31,8 +31,8 @@ public final class Constants {
         public static final double shootRPM = -4250;
         public static final PIDFGains gains = new PIDFGains(kP, 0, 0, kF, 0, 0);
         public static final int maxAccelerationSeconds = 1;
-        public static final float kVelocityConversionFactor = 600 / 2048; // 100ms per minute / upr
-                                                                          // (native velocity to RPM)
+        public static final double kVelocityConversionFactor = 600.0 / 2048; // 100ms per minute / upr
+                                                                             // (native velocity to RPM)
         public static final double kPeakOutput = 1;
     }
